+++ CodenameOne/src/com/codename1/io/gzip/InfCodes.java
@@ -529,7 +529,7 @@ final class InfCodes {
 
                 if ((e & 16) != 0) {
                     e &= 15;
-                    c = tp[tp_index_t_3 + 2] + ((int) b & inflate_mask[e]);
+                    c = tp[tp_index_t_3 + 2] + (b & inflate_mask[e]);
 
                     b >>= e;
                     k -= e;
